- created another component called IssueList - fixed indentation - add spacing between issues - transferred pagination logic in utils
