+++ src/components/Labels.js
@@ -2,13 +2,13 @@ import React from 'react';
 
 
 const Labels = ({ labels }) => {
-  let labelMap = labels.map((label, index) => {
-    return <span key={index} style={{ backgroundColor: "#" + label.color }} className={"label"}>{label.name}</span>
-})
+    let labelMap = labels.map((label, index) => {
+      return <span key={index} style={{ backgroundColor: "#" + label.color }} className={"label"}>{label.name}</span>
+  })
   return (
-    <React.Fragment>
-      {labelMap}
-    </React.Fragment>
+      <React.Fragment>
+          {labelMap}
+      </React.Fragment>
   )
 
 }
